cmd/snap-exec: move sources around
